pyproject.toml
[project]
name = "aiuc-core"
version = "0.1.0"
description = "Taxonomy-driven, multi-turn adversarial evaluation engine for AI agent safety testing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.71.0",
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=1.109.1",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
    "reportlab>=4.0.0",
    "rich>=13.9.4",
    "ruff>=0.14.1",
    "tqdm>=4.67.1",
    "verdict>=0.2.7",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "modal>=0.64.0",
    "fastcrud>=0.19.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "greenlet>=3.2.4",
    "absolufy-imports>=0.3.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
